Headphones now work if already plugged in at boot
If headphones were plugged in at boot, sound would not work. Now the headphones
are initialised as if they were plugged in manually. This seems to mostly work
around 50% of the time, whereas the other 50% results in the
"ERROR - max count exceeded" error in cs_8409_read_status_and_clear_interrupt.

I suspect an unwanted response is being queued and messing with the rest of the
driver. If I can clear it (or stop it from happening) then it may work better.

It's a naive approach to fixing the problem but a 50% success rate is better
than none.